+++ routers/router.go
@@ -1,22 +1,18 @@
 package routers
 
 import (
-	"fmt"
-	"log"
-
 	"net/http"
-	"os"
 
 	"github.com/JuliaKozachuk/BackChat/controllers"
 	"github.com/JuliaKozachuk/BackChat/migrations"
 
 	docs "github.com/JuliaKozachuk/BackChat/docs"
 	"github.com/gin-gonic/gin"
-	"github.com/joho/godotenv"
 	swaggerfiles "github.com/swaggo/files"
 	ginSwagger "github.com/swaggo/gin-swagger"
 )
 
+//@title BACKCHAT Api
 // @BasePath /api/v1
 
 // PingExample godoc
@@ -32,7 +28,7 @@ import (
 func Helloworld(g *gin.Context) {
 	g.JSON(http.StatusOK, "helloworld")
 }
-func Router() {
+func InitRouter() *gin.Engine {
 	route := gin.Default()
 
 	docs.SwaggerInfo.BasePath = "/api/v1"
@@ -45,38 +41,9 @@ func Router() {
 
 		}
 	}
-
-	migrations.ConnectDB(postgresUrl())
-	//migrations.Missing()
-
 	route.GET("/", func(context *gin.Context) {
 		context.JSON(http.StatusOK, gin.H{"message": migrations.Alex()})
-
 	})
-
 	route.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerfiles.Handler))
-	err := route.Run(":9888")
-	if err != nil {
-		panic("[Error] failed to start Gin server due to: " + err.Error())
-		return
-	}
-
-	// route.Run()
-
-}
-func postgresUrl() string {
-	err := godotenv.Load()
-	if err != nil {
-		log.Fatal("Error loading .env file")
-	}
-	host := os.Getenv("POSTGRES_HOST")
-	port := os.Getenv("POSTGRES_PORT")
-	user := os.Getenv("POSTGRES_USER")
-	dbname := os.Getenv("POSTGRES_DB")
-	password := os.Getenv("POSTGRES_PASSWORD")
-	sslmode := os.Getenv("POSTRES_SSLMODE")
-
-	postgres_data := fmt.Sprintf("host=%s port=%s user=%s dbname=%s password=%s sslmode=%s ", host, port, user, dbname, password, sslmode)
-
-	return postgres_data
+	return route
 }
